Fix `fyne release --os android` not respecting `ldflags` present in `GOFLAGS` enviroment variable
