+++ cmd/fyne/internal/mobile/build.go
@@ -236,8 +236,10 @@ func RunNewBuild(target, appID, icon, name, version string, build int, release, 
 	buildBundleID = appID
 	buildRelease = distribution
 	buildTags = tags
+	buildLdflags, _ = util.ExtractLdflagsFromGoFlags(os.Environ())
 	if release {
-		buildLdflags = "-w"
+		buildLdflags += " -w"
+		buildLdflags = strings.TrimSpace(buildLdflags)
 		buildTrimpath = true
 	}
